fix: long-running table lock if there is a migration
Suspect that adding a column holds an access exclusive lock on the table, even
though it doesn't re-write the entire table. So, to avoid ingest holding this
lock during a batch, the migrations have to be committed, and so have to then
be visible before the first batch.

Other visibility behaviour is possible - for example always ingesting into a
migration table and swapping with the live at the end. But this would require a
copy of all data in all migration circumstances.