feat: delete before first batch
This converts the Delete.ALL to Delete.BEFORE_FIRST_BATCH. In the
case of there being batches ingested, the behaviour is identical.
In the case of there being no batches, no data is deleted. This is
helpful for the cases where each batch contains the entire dataset
but if no batches after high watermark, we want to leave the data
alone.