build(deps): avoid axios vulnrability (again)
This copies the package-lock.json from stream-zip (changing the package name)
in an effort to avoid https://github.com/advisories/GHSA-wf5p-g6vw-rhxx

Not really sure why this had to be done - the package.json file is identical
with stream-zip's, so npm install of it should result of the same lock file.
But somehow it doesn't. Judging it not worth it to find out.