refactor: move is_upsert calculation close to where it's used
This should make it slightly easier to move to having multiple tables, because
the is_upsert value will be table-specific, and so we won't have to store an
array of is_upsert values.

Technically this will be a tiny bit less performant, because it will be
calculated on every batch. But if batches are so small that this is an issue,
it's likely there are other things that are much worse.