feat: support multiple tables
This adds support for ingesting to multiple tables in a single call to the
ingest function.

This works by having buffering incoming data, up to a maximum of 10k rows per
table. But it does it in a slightly fancy way, where

- The data is still streamed in for the first table using COPY FROM - it
  doesn't wait for a buffer to be filled.
- But during this, rows that appear in the stream for other tables are
  buffered.
- If a buffer hits 10k rows then the first COPY FROM is ended, and a new one
  begun, starting off with an empty of its 10k buffer, but then continues in a
  streaming way. And this one continues until the next table fills its buffer,
  and so on.

This means that the behaviour for a single table is exactly as now: no buffer
is used because it's all streamed straight in in a single COPY FROM. And the
stickiness also means that the number of separate COPY FROMs should be
minimised, no matter the order that the table/rows appear in the source stream.

Other alternatives were considered:

- Using separate connections. This would reduce memory use and be more
  streaming, but we would lose the transaction-ness of a single batch. This
  _is_ PostgreSQL after all, let's leverage what it gives us.
- Not using COPY FROM, but say separate INSERT statements. Would avoid
  buffering, but each INSERT would be slower.