feat: only GRANT when ingesting into a new table
Copying the SELECT permissions on each batch is semantically fine, but

- Unnecessary at best
- And at worst, can cause hanging and "tuple concurrently updated" errors when
  other clients try to do GRANT SELECT on the tables in question

This change makes it so the copy of SELECT privileges only happens when we
ingest the first batch into a new table, and never otherwise.

There is an existing test
test_migrate_add_column_not_at_end_permissions_preserved that test permissions
are preserved.