fix: save_comment method to use passed table variables
Issue identified when a single batch of data is ingested to multiple SQL tables, the high watermark is not written to the correct table (the first in the list of target tables) in order to be used in subsequent DAG runs. This is amended by using the schema and table variables as defined in the save_comment method. A test for this (test_multiple_tables_high_watermark) can be found in the test script.