fix: only commit once per batch, even when recreating views
This removes commits per view when recreating views, and restoring a commit
once per batch approach, so changes are only seen by other clients at the end
of a batch and not before.

This does mean the exclusive locks on all the views are accumulated, but I
think that is the least-bad option, because otherwise views will be visible in
an inconsistent state (/ without any data in some cases), albeit briefly.