fix: preservation of SELECT on views by view owner
This adds in the use of `acldefault` when finding existing SELECT privileges on
a view. This is so that SELECT privileges are not lost when views owned by
users other than doing the ingest are recreated.

Suspect this is still another change after this - to make sure ownership of
recreated views are preserved.