fix: preserve ownership of dependent views
This makes sure that ownership of dependent views when they are owned by
a user different to the one doing the ingest is maintained.

This includes a test of a related property: making sure that ingest into a
table works when there is a dependent view in a schema that is owned by a
different user, but this was passing before, so is added more as pro-active
regression test, since this change does involve changing things relating to
ownership.

NOTE: In systems that have a database trigger to maintain ownerships of objects
in schemas, such as Data Workspace, this isn't needed.