refactor: push skipping tables earlier in processing
This pushes skipping tables not in metadata earlier in the processing, and also
makes it so the low level ingest function just has an iterable of rows for a
table, rather than table,row pairs

This is a stepping stone to more ways of ingesting into the database, so each
has less responsibility, and also a more standard structure of data, i.e.
iterable of rows.