refactor: test multiple engine types/dialects each test run via parameterisation
This refactors tests, including how they are run in ci, and specifically:

- The engine type (aka dialect) is chosen by parameterisation, rather than what
  happens to be installed
- The versions installed are wherever possible enshrined in pyproject.toml via
  constraints on the python_version

This is done primarily to make it easier to test with more dialects (e.g. with
an ADBC dialect to ingest arrow table more efficiently), but it also makes the
ci config more understandable without the many excludes.