refactor: move injection of csv copy from stdin to earlier in processing
This moves the injection of the type of csv copy from stdin i.e. the one of
psycopg2 or the one for psycopg3, earlier in the processing, so core code of
ingestion doesn't need to pass any copy_from_stdin function.

This is a step towards allowing more way of ingesting data, and specifically
ones that don't need a copy from stdin function passed to them (for example
ingesting Arrow tables via an ADBC SQLAlchemy dialect).

This also tightens up what happens if modules are not installed but attempted
to be used - a KeyError should be raised right at the beginning of the ingest,
rather than attempted to use something that's None later (the KeyError could be
surfaced as a nicer error in future changes). This also groups the
psycopg2/psycopg related code closer together in the source file.