refactor: per-dialect as_string
This makes is so the as_string function, used to convert sql.SQL instances to
strings, can be defined per dialect.

This is to more easily support _other_ dialects, and specifically an ADBC
dialect that will have a slightly different as_string function.

Along the way, also changeing the ingest functon to not pass around variables
that are constant since close to the beginning of the ingest function to cut
noise. The exception is the insert_rows function, because this might end up
overwritable by client code, and I think it would be good to have the
connection passed to it.